Add data to books and game
